% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.mcmc.crr_fit}
\alias{as.mcmc.crr_fit}
\title{Convert to coda mcmc.list}
\usage{
\method{as.mcmc}{crr_fit}(x, ...)
}
\arguments{
\item{x}{A \code{crr_fit} object.}

\item{...}{Unused.}
}
\value{
A \link[coda:mcmc.list]{coda::mcmc.list}.
}
\description{
Post-warmup, thinned draws per chain.
}
