% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as_draws.crr_fit}
\alias{as_draws.crr_fit}
\title{Convert to posterior draws_array}
\usage{
\method{as_draws}{crr_fit}(x, ...)
}
\arguments{
\item{x}{A \code{crr_fit} object.}

\item{...}{Unused.}
}
\value{
A \code{posterior::draws_array} (requires the 'posterior' package).
}
\description{
Convert to posterior draws_array
}
