% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combreg-package.R
\docType{package}
\name{combreg-package}
\alias{combreg}
\alias{combreg-package}
\title{combreg: Bayesian Regression for Combinatorial Response Data}
\description{
Bayesian regression whose response data are integer-valued vectors subject to combinatorial constraints in the form of Ay<=b. Implements the Metropolis-Hastings-within-Gibbs sampler of Zheng et al. (2026) \doi{10.1080/01621459.2026.2714575}, an unconstrained probit baseline for comparison, benchmarking helpers, MCMC and regression diagnostics (effective sample sizes, split-Rhat, structured reports), plotting methods (trace, autocorrelation, violin, effective-sample-size, sampling-efficiency, residual heat map), and utilities for constraint validation (total unimodularity, feasibility) and data simulation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YuZh98/combreg}
  \item Report bugs at \url{https://github.com/YuZh98/combreg/issues}
}

}
\author{
\strong{Maintainer}: Hugh Zheng \email{hugh.stats@gmail.com} [copyright holder]

}
\keyword{internal}
