% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{crr_prior}
\alias{crr_prior}
\title{Prior specification for crr models}
\usage{
crr_prior(family = "gaussian", sd = 1)
}
\arguments{
\item{family}{Prior family; only \code{"gaussian"} is implemented.}

\item{sd}{Prior standard deviation (scalar, positive).}
}
\value{
An object of class \code{crr_prior}.
}
\description{
Currently a single family: independent Gaussian priors
\eqn{\beta_{kj} \sim N(0, sd^2)} on all regression coefficients.
}
\examples{
crr_prior()               # N(0, 1)
crr_prior(sd = 10)        # weakly informative

}
