% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{residuals.crr_fit}
\alias{residuals.crr_fit}
\title{Response residuals}
\usage{
\method{residuals}{crr_fit}(object, ...)
}
\arguments{
\item{object}{A \code{crr_fit} object.}

\item{...}{Unused.}
}
\value{
Integer matrix (\code{n} x \code{d}).
}
\description{
Training residuals \code{Y - fitted(object)}, entries in \{-1, 0, 1\}.
}
