% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.crr_fit}
\alias{coef.crr_fit}
\title{Posterior mean coefficients}
\usage{
\method{coef}{crr_fit}(object, ...)
}
\arguments{
\item{object}{A \code{crr_fit} object.}

\item{...}{Unused.}
}
\value{
Posterior mean coefficient matrix (\code{p} x \code{d}).
}
\description{
Posterior mean coefficients
}
