% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combreg-package.R
\docType{package}
\name{combreg-package}
\alias{combreg}
\alias{combreg-package}
\title{combreg: Bayesian Regression for Combinatorial Response Data}
\description{
Bayesian posterior sampling for regression models whose response is an integer-valued vector constrained to a polytope defined by a totally unimodular system, as in matching and assignment problems. Implements the Metropolis-Hastings-within-Gibbs sampler with dual-variable augmentation of Zheng et al., with hit-and-run updates for the dual certificates accelerated in 'C++', an unconstrained probit baseline for comparison, and utilities for constraint validation (total unimodularity, feasibility) and data simulation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YuZh98/combreg}
  \item Report bugs at \url{https://github.com/YuZh98/combreg/issues}
}

}
\author{
\strong{Maintainer}: Yu Zheng \email{seanzhengyu1@gmail.com}

}
\keyword{internal}
